Add onTap handler to ListTile in CreatePostView
This commit introduces an onTap handler to the ListTile widget within the
CreatePostView of the post feature. The addition of this handler is intended
to enable user interaction with the ListTile, specifically for selecting
photos, as indicated by the accompanying icon. This change is part of the
effort to improve the user experience by making the UI elements more
interactive and responsive.