Fix Typo and Enhance Post Item Widget Functionality
Corrected a typo in the category name from "Laptobs" to "Laptops" in the main view body widget. Improved the post item widget by conditionally displaying the "Remove Post" option only for the post owner, and wrapping the CachedNetworkImage with a ClipRRect for better visual appearance. Additionally, ensured that the getAllProducts function is awaited when selecting the "All Products" category. Removed the unnecessary orderBy clause from the Firestore query in the GetPostsCubit to streamline the query logic.